+++ src/main/java/com/raoulvdberge/refinedstorage/api/util/IQuantityFormatter.java
@@ -17,6 +17,18 @@ public interface IQuantityFormatter {
     String formatWithUnits(int qty);
 
     /**
+     * Formats a quantity as they are formatted in the Grid.
+     * Formatted as following: "####0.#".
+     * <p>
+     * If the quantity is equal to or bigger than 1000 it will be displayed as the quantity divided by 1000 (without any decimals) and a "K" appended.
+     * If the quantity is equal to or bigger than 1000000 it will be displayed as the quantity divided by 1000000 (without any decimals) and a "M" appended.
+     *
+     * @param qty the quantity
+     * @return the formatted quantity
+     */
+    String formatWithUnits(long qty);
+
+    /**
      * Formats a quantity as they are formatted on the disk tooltips.
      * Formatted as following: "#,###".
      *
@@ -24,6 +36,15 @@ public interface IQuantityFormatter {
      * @return the formatted quantity
      */
     String format(int qty);
+
+    /**
+     * Formats a quantity as they are formatted on the disk tooltips.
+     * Formatted as following: "#,###".
+     *
+     * @param qty the quantity
+     * @return the formatted quantity
+     */
+    String format(long qty);
 
     /**
      * Divides quantity by 1000 and appends "B".